pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kws-2026-submission"
version = "0.1.0"
description = "Dual-task KWS submission with KeywordMamba + MHAtt-CRNN"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "torch>=2.5,<2.7",
  "torchaudio>=2.5,<2.7",
  "mambapy>=1.2.0",
  "librosa>=0.11,<0.12",
  "sounddevice>=0.5,<0.6",
  "numpy>=1.26,<3",
  "pandas>=2.2,<3",
  "scikit-learn>=1.5,<2",
  "PyYAML>=6.0,<7",
  "huggingface_hub>=0.30,<1",
  "transformers>=4.51,<5",
  "tqdm>=4.66,<5",
  "matplotlib>=3.9,<4",
  "tensorboard>=2.18,<3",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
